requirements.txt
django==3.0.5
psycopg2>=2.8.4
django-cors-headers==3.2.1
djangorestframework>=3.11.0
django-filter>=2.2.0
djangorestframework-filters==1.0.0.dev0
djangorestframework-simplejwt==4.6.0
pylint>=2.5.2
pylint-django>=2.0.15
channels==2.4.0
channels-redis
numpy
scipy
pandas
plotly==4.8.0
matplotlib
git+https://github.com/ibis-inria/wellFARE.git
cvxopt
django-pandas
decorator
cycler==0.10.0
django-extensions==2.1.3
et-xmlfile==1.0.1
jdcal==1.4
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.0
openpyxl==3.0.4 #2.5.9
Pillow==6.2.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.5
six==1.12.0
statsmodels==0.10.2
ipython==7.11.1
httptools
uvloop
uvicorn[standard]
gunicorn
daphne
twisted==20.3.0
